requirements.txt
streamlit>=1.31.0
opencv-python>=4.9.0.80
numpy>=1.26.0
Pillow>=10.0.0
requests>=2.31.0
python-uuid>=1.30
uuid>=1.30